third_party/blink/web_tests/platform/linux/virtual/sanitizer-api/external/wpt/sanitizer-api/sanitizer-get.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Sorting of attributes
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of removeAttributes
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of elements
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of removeElements
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of replaceWithChildrenElements
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of element's attributes
  assert_false: Either elements or a removeElements, but not both expected false got true
[FAIL] Sorting of element's removeAttributes
  assert_false: Either elements or a removeElements, but not both expected false got true
Harness: the test ran to completion.
